+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added public function `Get-WebCert`.
+
 ### Fixed
 
 - Fixed missing comment help block in `Get-ADHostAudit`